Refactor hash algorithms into private headers
