+++ src/private/hash_algorithm.h
